Move Md5CheckSumClient to aws module
